*
I wrote the code however it came to including his library which brought
many many troubles with it! Why give this lecture if including this
library is not easy ?? I had to find the library seperately which inside
required additional .h docs which were not compatible and needed
conversion and which after this convertion I could not run at all due to
an undefined variable inside the conf.h. moreover there is the problem
with this arc32 architecture! I dont like this:/