%
I completed the project so far until the point that I can charge the
battery with the maximum power point when solar panel is exposed
to sun at a good day light or at least can provide the battery a voltage
larger than the battery's current voltage. The system control can
tolarate very small changes in the light. However since the light
conditions were not perfect under the circumstances I studied, some
slight
changes caused the control to get out of control. The addition in the
contorl code is to make it work better keeping the previous P datas so
to make
sure that it does not fluctuate too much . Yet it occurred thatit was
not necessary since I realized that the duty cycle does not increase in
floating
nuber but as an integer. I did not delete that part since I spent quite
a lot of time on it. I also tried the code to restart itself when this
jiggling occurs during changes in the daylight which would mimic the
cloud covering the sun for instance but there is not a speial command in
Arduino which will restart it when the system is resetted (I have
already resetted the system). Yet, I could recharge a battery from 2.75
to
3.7 V in the provided time.